Add proper behaviour of ground robots
new_arena.launch is the launch file for the new arena
The new arena contain only the ground bots and patrol bots, both of whose behaviours have been added
Only work left on the behavious for ground robot is the response to quad hovering over it
ContactPlugin is the name of the new plugin implemented, which detects collisions of the bots,
while groundbot.cpp plugin determines theh behaviour based on collision information from the
ContactPlugin. This information is sent over through a ros topic.