Add the model for top-sensor for ground bots
To implement a sensor to detect if the quad touches a ground bot's top surface,
a simple cylinder object will be placed on top of the ground bot with a contact
sensor. This commit brings the modification to each ground bot to have a cylinder
with a longer(higher) collision dimension than the visual object.