Add top sensor behaviour, fixed #4 and #5
Fixed issues: #4[1] and #5[2] :)
Top contact sensor trigger is read by the flag '2' in /contact
topic for the particular ground bot.

The issue #4[2], where the groudbots used to turn before the
timeout even without any collision event was due to a condition
that monitored deviation between target and current heading as
the former get changed due to error in forward movement. So, it
was fixed by adding another bool is_turning which specifically
tells when the ground bot needs to turn.

[1]: https://github.com/quadrotor-IITKgp/quad_simulator/issues/4
[2]: https://github.com/quadrotor-IITKgp/quad_simulator/issues/5